fix(firebase): fix operators and import from separate file for clarity
