+++ packages/cerebral-provider-firebase/src/index.js
@@ -28,34 +28,6 @@ import sendPasswordResetEmail from './sendPasswordResetEmail'
 import linkWithGithub from './linkWithGithub'
 import {setOnDisconnect, cancelOnDisconnect} from './onDisconnect'
 
-export {default as createUserWithEmailAndPassword} from './factories/createUserWithEmailAndPassword'
-export {default as getUser} from './factories/getUser'
-export {default as off} from './factories/off'
-export {default as onChildAdded} from './factories/onChildAdded'
-export {default as onChildChanged} from './factories/onChildChanged'
-export {default as onChildRemoved} from './factories/onChildRemoved'
-export {default as onValue} from './factories/onValue'
-export {default as signInAnonymously} from './factories/signInAnonymously'
-export {default as signInWithEmailAndPassword} from './factories/signInWithEmailAndPassword'
-export {default as signInWithFacebook} from './factories/signInWithFacebook'
-export {default as signInWithGoogle} from './factories/signInWithGoogle'
-export {default as signInWithGithub} from './factories/signInWithGithub'
-export {default as signOut} from './factories/signOut'
-export {default as task} from './factories/task'
-export {default as value} from './factories/value'
-export {default as set} from './factories/set'
-export {default as update} from './factories/update'
-export {default as push} from './factories/push'
-export {default as put} from './factories/put'
-export {default as delete} from './factories/delete'
-export {default as remove} from './factories/remove'
-export {default as transaction} from './factories/transaction'
-export {default as deleteUser} from './factories/deleteUser'
-export {default as sendPasswordResetEmail} from './factories/sendPasswordResetEmail'
-export {default as linkWithGithub} from './factories/linkWithGithub'
-export {default as setOnDisconnect} from './factories/setOnDisconnect'
-export {default as cancelOnDisconnect} from './factories/cancelOnDisconnect'
-
 export default function FirebaseProviderFactory (options = { payload: {} }) {
   firebase.initializeApp(options.config)
