+++ packages/cerebral-provider-firebase/src/operators/cancelOnDisconnect.js
