+++ packages/cerebral-provider-firebase/src/operators/createUserWithEmailAndPassword.js
