+++ packages/cerebral-provider-firebase/src/operators/delete.js
