+++ packages/cerebral-provider-firebase/src/operators/deleteUser.js
