+++ packages/cerebral-provider-firebase/src/operators/getUser.js
