+++ packages/cerebral-provider-firebase/src/operators/linkWithFacebook.js
