+++ packages/cerebral-provider-firebase/src/operators/linkWithGithub.js
