+++ packages/cerebral-provider-firebase/src/operators/linkWithGoogle.js
