+++ packages/cerebral-provider-firebase/src/operators/off.js
@@ -1,6 +1,6 @@
 function offFactory (path, event, signal) {
   function off ({firebase, resolve}) {
-    firebase.off(resolve.value(path), resolve.value(event), resolve.value(signal))
+    firebase.off(resolve.value(path), resolve.value(event), resolve.path(signal))
   }
 
   return off
