+++ packages/cerebral-provider-firebase/src/operators/onChildAdded.js
@@ -2,7 +2,7 @@ import {convertObjectWithTemplates} from './utils'
 
 function onChildAddedFactory (path, signal, options = {}) {
   function onChildAdded ({firebase, resolve}) {
-    firebase.onChildAdded(resolve.value(path), resolve.value(signal), convertObjectWithTemplates(options, resolve))
+    firebase.onChildAdded(resolve.value(path), resolve.path(signal), convertObjectWithTemplates(options, resolve))
   }
 
   return onChildAdded
