+++ packages/cerebral-provider-firebase/src/operators/onChildChanged.js
@@ -2,7 +2,7 @@ import {convertObjectWithTemplates} from './utils'
 
 function onChildChangedFactory (path, signal, options = {}) {
   function onChildChanged ({firebase, resolve}) {
-    firebase.onChildChanged(resolve.value(path), resolve.value(signal), convertObjectWithTemplates(options, resolve))
+    firebase.onChildChanged(resolve.value(path), resolve.path(signal), convertObjectWithTemplates(options, resolve))
   }
 
   return onChildChanged
