+++ packages/cerebral-provider-firebase/src/operators/push.js
