+++ packages/cerebral-provider-firebase/src/operators/remove.js
