+++ packages/cerebral-provider-firebase/src/operators/sendPasswordResetEmail.js
