+++ packages/cerebral-provider-firebase/src/operators/set.js
