+++ packages/cerebral-provider-firebase/src/operators/setOnDisconnect.js
