+++ packages/cerebral-provider-firebase/src/operators/signInAnonymously.js
