+++ packages/cerebral-provider-firebase/src/operators/signInWithEmailAndPassword.js
