+++ packages/cerebral-provider-firebase/src/operators/signInWithFacebook.js
