+++ packages/cerebral-provider-firebase/src/operators/signInWithGithub.js
