+++ packages/cerebral-provider-firebase/src/operators/signInWithGoogle.js
