+++ packages/cerebral-provider-firebase/src/operators/signOut.js
