+++ packages/cerebral-provider-firebase/src/operators/task.js
