+++ packages/cerebral-provider-firebase/src/operators/transaction.js
