+++ packages/cerebral-provider-firebase/src/operators/update.js
