+++ packages/cerebral-provider-firebase/src/operators/utils.js
