+++ packages/cerebral-provider-firebase/src/operators/value.js
